untitled

+++ main.c
@@ -6,6 +6,8 @@
 #include <sys/wait.h>
 #include "run/main.h"
 
+char *rootworkdir;
+
 void loadcwd(char *rootworkdir)
 {  
     char path[NBUF]; snprintf(path, NBUF, "%s/cwd.txt", rootworkdir);
@@ -20,44 +22,18 @@ void loadcwd(char *rootworkdir)
     }
 }
 
-typedef struct {                       
+typedef struct {  
+  int isfirstcmd;                     
   int syssh;                                                                                   
   char *rootworkdir;                                                                         
   int setemptyline;
-  char *lastline;  
+  char *lastline; int lastlinen;
   char *linebuf;
   char *line;
   char *cwd; 
 
   FILE *srcfile;                                                                       
 } maincontext_t;
-
-void maingetline(maincontext_t *ctxt) 
-{
-    ctxt->linebuf[0] = 0;
-
-    int displaynumber = 1;
-    while loop {
-	    char *prtline = 0; size_t prtlinen = 0; ssize_t r;
-	    r = getline(&prtline, &prtlinen, stdin);
-    	throwngerrno(r);
-     	if (strlen(ctxt->linebuf) + strlen(prtline) < NBUF) {
-    		strcat(ctxt->linebuf + strlen(ctxt->linebuf), prtline);
-    	} else {
-    		printf("Buffer overflow.\n");
-    		exit(1);
-    	}
-	
-       free(prtline);
-       if (strlen(prtline) < 2 
-       || (prtline[strlen(prtline) - 2] != '\\'))
-  	       break;
-           printf("%d ", ++displaynumber);
-       }
-    
-    printf("\n");
-}
-
 
 char *ctxtmwf(malloclist_t *malloclist, maincontext_t *ctxt, char *path)
 {
@@ -72,9 +48,10 @@ void mainprintsubmainbegin(maincontext_t *ctxt)
     	fprintf(ctxt->srcfile, "#include \"main.h\"\n");
 	fprintf(ctxt->srcfile, "\n");
 	fprintf(ctxt->srcfile, "char *lastcmd = \"%s\";\n", 
-	    escapecstr(ctxt->lastline, linebuf_tn));
+	    escapecstr(ctxt->lastline, ctxt->lastlinen));
 	fprintf(ctxt->srcfile, "int syssh = %d;\n", ctxt->syssh);
 	fprintf(ctxt->srcfile, "char *rootworkdir = \"%s\";", ctxt->rootworkdir);
+	fprintf(ctxt->srcfile, "int isfirstcmd = %d;", ctxt->isfirstcmd);
 	fprintf(ctxt->srcfile, "\n");
 	fprintf(ctxt->srcfile, "int main(int argc, char **argv) {\n");
 	fprintf(ctxt->srcfile,     "MAIN_BEGIN\n");	
@@ -95,13 +72,28 @@ void maincdcase(maincontext_t *ctxt)
 
 		ctxt->line = ctxt->line + strlen(ctxt->line);
 	}	
+
+	if (0 == strncmp(ctxt->line, "# cdroot ", 9)) {
+		ctxt->line += 9;
+		ctxt->line[strlen(ctxt->line) - 1] = 0;
+			
+		fputsclose
+		(
+		    mf("%s/\n", ctxt->line), 
+		    fopen(ctxtmwf("cwd.txt"), "w")
+		);
+		loadcwd(ctxt->rootworkdir);
+
+		ctxt->line = ctxt->line + strlen(ctxt->line);
+	}	
 }
 
-void mainfixinputline(maincontext_t *ctxt)
+void mainremoveunescnl(char *line)
 {
-	for (unsigned long i = 0; i < strlen(ctxt->line); ++i) {
-		if ((ctxt->line[i] == '\n' || ctxt->line[i] == (char)10) && ctxt->line[i-1] != '\\')
-		    ctxt->line[i] = ' ';
+	for (unsigned long i = 1; i < strlen(line); ++i) {
+		if ((line[i] == '\n' || line[i] == (char)10))
+			if ((line[i-1] != '\\'))
+		    	line[i] = ' ';
 	}    	
 }
 
@@ -115,12 +107,39 @@ void mainprintsubmainend(maincontext_t *ctxt)
 
 int main(int argc, char **argv) {
   maincontext_t *ctxt = mallocadd(sizeof(maincontext_t));
-  
-  printf("[press enter] ");
 
+  linebuf_t runopt;
+
+  int defstrchrnul = 0;
+
+  /*
+  while loop {
+	//printf("prooted? [y] ");
+	fflush(stdout);
+	//char *answer = mgetescline("");
+
+  	if (!strncmp(answer, "y", 1)) {
+  	  defstrchrnul = 0;
+      break;
+  	}
+  	if (!strncmp(answer, "n", 1)) {
+      defstrchrnul = 1;
+      break;
+  	}
+  	defstrchrnul = 0;
+  	break;
+  }
+  */
+
+  fflush(stdout);
+  printf("[press enter] ");
+  fflush(stdout);
+
+  ctxt->isfirstcmd = 1;
   ctxt->syssh = 0;
   
   ctxt->rootworkdir = argv[1];
+  rootworkdir = ctxt->rootworkdir;
 
   if (ctxt->rootworkdir[strlen(ctxt->rootworkdir) - 1] == '/') {
   	ctxt->rootworkdir[strlen(ctxt->rootworkdir) - 1] = 0;
@@ -128,10 +147,13 @@ int main(int argc, char **argv) {
   
   if (argc < 2) { printf("%s: %s %s%s\n", "Usage", argv[0], "<workdir>", ""); return 1; }
 
-  ctxt->lastline = mallocaddlinebuf;
+  ctxt->lastline = mallocadd(10 * linebuf_tn);
+  ctxt->lastlinen = 110 * linebuf_tn;
   ctxt->cwd = mallocaddlinebuf;
   ctxt->setemptyline = 0;
-  ctxt->linebuf = mallocaddlinebuf;
+  ctxt->linebuf = mallocaddpagebuf;
+
+  int nextbackup = 10;
 
   while loop {
     loadcwd(ctxt->rootworkdir);
@@ -142,8 +164,20 @@ int main(int argc, char **argv) {
     //ctxt->setemptyline = 0;
 
     if (!ctxt->setemptyline)
-    {
-    	maingetline(ctxt);
+    loctag {
+        ctxt->linebuf[0] = 0;
+
+        char page[pagebuf_tn];
+        page[0] = 0;
+		fgetallsclose(page, pagebuf_tn - 1, fopen(ROOTC("/insert.c"), "r"));
+
+        // add fskipwh here
+    iff !strlen(page)
+    thn	{    strncat(ctxt->linebuf, mgetescline(mf("%s%%02d ", GLOBAL_INDENT)), pagebuf_tn - 1);ln;}
+    els {    strncat(ctxt->linebuf, page,                                       pagebuf_tn - 1);
+             fputsclose("", fopen(ROOTC("/insert.c"), "w"));
+             printf("<insert.c>\n\n");
+        }
     }
     else {
     	ctxt->linebuf[0] = '\n';
@@ -152,13 +186,60 @@ int main(int argc, char **argv) {
     }
     ctxt->line = ctxt->linebuf;
 
-    in(ctxt->line, fputsclose, fopen(ctxtmwf("log.txt"), "a"));
-
-    if (!ctxt->syssh && 0 == strncmp(ctxt->line, "# exit", 6)) {
+    if (0 == strncmp(ctxt->line, "# exit", 6)) 
+    {
     	exit(0);
     }
 
-    mainfixinputline(ctxt);
+    if (0 == strncmp(ctxt->line, "# restart", 9)) 
+    {
+    	exit(5);
+    }
+
+	char *cmdprefix = "# runopt ";
+    if (0 == strncmp(ctxt->line, cmdprefix, strlen(cmdprefix)))
+    {
+    	ctxt->line += strlen(cmdprefix);	
+   		ctxt->line[strlen(ctxt->line) - 1] = 0;
+   			
+		strncpy(runopt, ctxt->line, linebuf_tn - 1);
+		runopt[linebuf_tn - 1] = 0;
+   
+   		ctxt->line = ctxt->line + strlen(ctxt->line);
+    }
+
+    in(ctxt->line, fputsclose, fopen(ctxtmwf("log.txt"), "a"));
+
+    mainremoveunescnl(ctxt->line);
+
+
+    if (0 == strncmp(ctxt->line, "# 0x", 4)) {
+        printf("unimplemented");ln;
+
+    } if (0) {
+		int i = strtol(ctxt->line + 4, NULL, 16);
+
+    	ctxt->linebuf[0] = 0;
+    	ctxt->line = ctxt->linebuf;
+
+		while loop {
+			linebuf_t cmd;
+			cmdlogat(i, &cmd);
+	        int linebuflen = strlen(ctxt->linebuf); 
+	        if (linebuflen + strlen(cmd) + 4 < pagebuf_tn) {
+				
+				strncpy(&ctxt->linebuf[linebuflen], cmd, strlen(cmd) + 1);
+			} else {
+				printf("Error at %s:%d", __FILE__, __LINE__);ln;
+			}
+			
+			if (!strncmp(&cmd[strlen(cmd) - 2], "\\\n", 2)) {
+				++i;
+			} else {
+				break;
+			}
+		}
+    }
 
     mainprintsubmainbegin(ctxt);
 	maincdcase(ctxt);
@@ -166,14 +247,17 @@ int main(int argc, char **argv) {
 
     char quiet[] = " 2> /dev/null > /dev/null";
 
+    sh, "mv %s %s %s", 
+        mf("-f %s/%s", ctxt->rootworkdir, "a.out"), 
+        mf("%s/%s", ctxt->rootworkdir, "b.out"), 
+        QUIET, endsh;
+
 	// TODO write back to ? : && || syntax
 	(sh, "rm -f %s", ctxtmwf("a.out"), endsh);
-	if (sh, "gcc %s -o%s", ctxtmwf("main.c"), ctxtmwf("a.out"), endsh) {
+	if (sh, "gcc %s %s -o%s", defstrchrnul ? "-D DEFSTRCHRNUL" : "", ctxtmwf("main.c"), ctxtmwf("a.out"), endsh) {
 		if (strncmp(ctxt->linebuf, " ", 2)) { ctxt->setemptyline = 1; } else {
-		    printf(s, "A1");ln;breakpt;
 			if (sh, "stat %s %s", mf("%s%s", ctxt->rootworkdir, ".err"), quiet, endsh)
 			{
-				printf(s, "A2");ln;breakpt;
 				if (sh, "mv %s %s", 
 				     ctxt->rootworkdir, mf("%s%s", ctxt->rootworkdir, ".err", endsh)
 				) { printf("Error at %s:%d\n", __FILE__, __LINE__); } else {
@@ -185,17 +269,24 @@ int main(int argc, char **argv) {
 				printf(s, mf("%s%s %s", ctxt->rootworkdir, ".err", "already exists"));ln;
 		}
 	} else {
-    	(sh, ctxtmwf("a.out"), endsh),
-    	(sh, "rm -r %s",    mf("%s%s", ctxt->rootworkdir, ".bak"), 
-    	                  /*mf("%s%s", ctxt->rootworkdir, ".bak.bak"),*/ endsh),
-    	(sh, "mv %s %s",    mf("%s%s", ctxt->rootworkdir, ".tmp.bak"),
-    	                    mf("%s%s", ctxt->rootworkdir, ".bak"), endsh),
-	    (sh, "cp -r %s %s", mf("%s%s", ctxt->rootworkdir, ""),
-	                        mf("%s%s", ctxt->rootworkdir, ".tmp.bak"), endsh);
-	  //(sh, "mv %s %s%s",  mf("%s%s", ctxt->rootworkdir, ".bak.bak"),
-	  //                    mf("%s%s", ctxt->rootworkdir, ".bak"), "" , endsh);
+    	(sh, "%s %s %s", runopt, ctxtmwf("a.out"), argc > 2 ? argv[2] : "", endsh);
+
+    	if (!nextbackup--)
+    	{
+    		nextbackup = 10,
+            (sh, "rm -r %s",    mf("%s%s", ctxt->rootworkdir, ".bak"), 
+    		                  /*mf("%s%s", ctxt->rootworkdir, ".bak.bak"),*/ endsh),
+    		(sh, "mv %s %s",    mf("%s%s", ctxt->rootworkdir, ".tmp.bak"),
+    		                    mf("%s%s", ctxt->rootworkdir, ".bak"), endsh),
+	        (sh, "cp -r %s %s", mf("%s%s", ctxt->rootworkdir, ""),
+	                            mf("%s%s", ctxt->rootworkdir, ".tmp.bak"), endsh);
+	      //(sh, "mv %s %s%s",  mf("%s%s", ctxt->rootworkdir, ".bak.bak"),
+	      //                    mf("%s%s", ctxt->rootworkdir, ".bak"), "" , endsh);
+		
+    	}
 	}
 	
 	strcpy(ctxt->lastline, ctxt->linebuf);
+	ctxt->isfirstcmd = 0;
   } 
 }
